docs: start migration guide for upcoming v0.8
