+++ .changeset/slimy-wasps-push.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-ai-sdk": patch
+---
+
+feat: RSCDisplay
