+++ packages/react-ai-sdk/src/rsc/index.ts
@@ -2,3 +2,4 @@ export { useVercelRSCRuntime } from "./useVercelRSCRuntime";
 export { getVercelRSCMessage } from "./getVercelRSCMessage";
 export type { VercelRSCAdapter } from "./VercelRSCAdapter";
 export type { VercelRSCMessage } from "./VercelRSCMessage";
+export { RSCDisplay } from "./RSCDisplay";
