+++ packages/react-ai-sdk/src/rsc/utils/RSCThreadExtras.tsx
@@ -1,0 +1,11 @@
+"use client";
+import { VercelRSCMessage } from "../VercelRSCMessage";
+
+export const symbolInternalRSCExtras = Symbol("internal-rsc-extras");
+export type RSCThreadExtras =
+  | {
+      [symbolInternalRSCExtras]?: {
+        convertFn: (message: any) => VercelRSCMessage;
+      };
+    }
+  | undefined;
